refactor: split up Internal.hpp
